Consolidate LockfileDependency.ts and LockfileEntry.ts into LfxGraph.ts
