+++ apps/lockfile-explorer-web/src/parsing/LfxGraph.ts
@@ -1,7 +1,48 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import type { LockfileDependency } from './LockfileDependency';
+export enum DependencyKind {
+  DEPENDENCY,
+  DEV_DEPENDENCY,
+  PEER_DEPENDENCY
+}
+
+/**
+ * Represents a dependency listed under a LockfileEntry
+ *
+ * @remarks
+ * Each dependency listed under a package in the lockfile should have a separate entry. These Dependencies
+ * will link to the "containingEntry", which is the LockfileEntry that specified this dependency.
+ * The "resolvedEntry" field is the corresponding LockfileEntry for this dependency, as all dependencies also have
+ * their own entries in the pnpm lockfile.
+ */
+export class LockfileDependency {
+  public name: string;
+  public version: string;
+  public entryId: string = '';
+  public dependencyType: DependencyKind;
+  public containingEntry: LockfileEntry;
+
+  public resolvedEntry: LockfileEntry | undefined;
+
+  public peerDependencyMeta: {
+    name?: string;
+    version?: string;
+    optional?: boolean;
+  } = {};
+
+  public constructor(options: {
+    name: string;
+    version: string;
+    dependencyType: DependencyKind;
+    containingEntry: LockfileEntry;
+  }) {
+    this.name = options.name;
+    this.version = options.version;
+    this.dependencyType = options.dependencyType;
+    this.containingEntry = options.containingEntry;
+  }
+}
 
 export enum LockfileEntryFilter {
   Project,
@@ -72,4 +113,8 @@ export class LockfileEntry {
   public constructor(kind: LockfileEntryFilter) {
     this.kind = kind;
   }
+}
+
+export class LfxGraph {
+  public entries: LockfileEntry[] = [];
 }
